refactor: move slider content and state to presentation layer
